backend/agents/software_engineering_team/architect_agents/requirements.txt
strands-agents>=1.35,<2.0
strands-agents-tools>=0.4,<1.0
ollama>=0.6,<1.0
boto3>=1.42,<2.0
bedrock-agentcore>=1.6,<2.0
